+++ features/Milestones/index.ts
@@ -1,4 +1,5 @@
 export * from './constants';
 export * from './data';
 export * from './Page';
+export * from './Line';
 export type * from './types';
